refacotor: url 구조 변경
