+++ comment/urls.py
@@ -2,6 +2,5 @@ from django.urls import path
 from .views import *
 
 urlpatterns = [
-    path('<int:board_id>', CommentView.as_view(), name='comment_list_create'),
-    path('<int:board_id>/<int:comment_id>', CommentDetailView.as_view(), name='comment_datail'),
+    path('<int:comment_id>', CommentDetailView.as_view(), name='comment_datail'),
 ]
