+++ post/views.py
@@ -1,27 +1,27 @@
 from rest_framework.views import APIView
 from rest_framework.response import Response
 from rest_framework import status, permissions
-from .serializers import PostCreateSerializer, PostSerializer
-from .models import Post
+from rest_framework.generics import get_object_or_404
 from rest_framework_simplejwt.authentication import JWTAuthentication
 from drf_yasg.utils import swagger_auto_schema
 from drf_yasg import openapi
-from rest_framework.generics import get_object_or_404
+
+from .models import Post
+from comment.serializers import CommentCreateSerializer, CommentSerializer
+from .serializers import PostCreateSerializer, PostSerializer
+
 
 class PostView(APIView):
     authentication_classes = [JWTAuthentication]
 
     def get_permissions(self):
-        # GET 요청은 비로그인 사용자도 허용, POST는 로그인 필요
         if self.request.method == 'GET':
             return [permissions.AllowAny()]
         return [permissions.IsAuthenticated()]
 
     @swagger_auto_schema(
         operation_summary="게시글 목록 조회",
-        responses={
-            200: openapi.Response(description="조회 성공", schema=PostSerializer(many=True))
-        }
+        responses={200: openapi.Response(description="조회 성공", schema=PostSerializer(many=True))}
     )
     def get(self, request):
         posts = Post.objects.all().order_by('-created_at')
@@ -41,9 +41,9 @@ class PostView(APIView):
         serializer = PostCreateSerializer(data=request.data, context={'request': request})
         if serializer.is_valid():
             post = serializer.save()
-            response_data = PostSerializer(post).data
-            return Response(response_data, status=status.HTTP_201_CREATED)
+            return Response(PostSerializer(post).data, status=status.HTTP_201_CREATED)
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+
 
 class PostDetailView(APIView):
     authentication_classes = [JWTAuthentication]
@@ -70,15 +70,13 @@ class PostDetailView(APIView):
             400: "잘못된 요청입니다.",
             401: "인증되지 않았습니다.",
             403: "작성자가 아닙니다.",
-            404: "게시글을 찾을 수 없습니다.",
+            404: "게시글을 찾을 수 없습니다."
         }
     )
     def put(self, request, post_id):
         post = get_object_or_404(Post, pk=post_id)
-
         if post.user_id != request.user:
             return Response({"detail": "수정 권한이 없습니다."}, status=status.HTTP_403_FORBIDDEN)
-
         serializer = PostCreateSerializer(post, data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -91,14 +89,34 @@ class PostDetailView(APIView):
             204: "삭제 성공",
             401: "인증되지 않았습니다.",
             403: "작성자가 아닙니다.",
-            404: "게시글을 찾을 수 없습니다.",
+            404: "게시글을 찾을 수 없습니다."
         }
     )
     def delete(self, request, post_id):
         post = get_object_or_404(Post, pk=post_id)
-
         if post.user_id != request.user:
             return Response({"detail": "삭제 권한이 없습니다."}, status=status.HTTP_403_FORBIDDEN)
-
         post.delete()
         return Response({"detail": "삭제 완료"}, status=status.HTTP_204_NO_CONTENT)
+
+
+class CommentView(APIView):
+    permission_classes = [permissions.IsAuthenticatedOrReadOnly]
+
+    @swagger_auto_schema(
+        operation_summary="댓글 생성",
+        request_body=CommentCreateSerializer,
+        responses={
+            201: openapi.Response(description="생성 성공", schema=CommentSerializer),
+            400: "잘못된 요청입니다.",
+            401: "인증되지 않았습니다.",
+            404: "게시글을 찾을 수 없습니다."
+        }
+    )
+    def post(self, request, post_id):  # post_id로 수정
+        post = get_object_or_404(Post, pk=post_id)
+        serializer = CommentCreateSerializer(data=request.data, context={'request': request, 'post': post})
+        if serializer.is_valid():
+            comment = serializer.save()
+            return Response(CommentSerializer(comment).data, status=status.HTTP_201_CREATED)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
